+++ README.md
@@ -11,7 +11,7 @@ The repository now separates the main components for clarity:
 ## Completed Features
 
 - **Python bootstrap interpreter** with a REPL, basic arithmetic, variables and functions. The environment now provides built-in list utilities `null?`, `length`, `map` and `filter`.
-- **Self-hosted evaluator** written in Lisp loaded by `run.py` via `(import ...)`.
+- **Self-hosted evaluator** written in Lisp loaded by `run_hosted.py` via `(import ...)`.
 - Lisp features implemented in Lisp:
   - `cond` form and `define-macro` for simple macros.
   - List utilities: `null?`, `length`, `map` and `filter`.
