+++ lispfun/__main__.py
@@ -1,4 +1,4 @@
-from .run import main
+from run_hosted import main
 
 if __name__ == "__main__":
     main()
