+++ lispfun/bootstrap/tests/test_import.py
@@ -3,7 +3,7 @@ import sys
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", "..")))
 
 from lispfun.interpreter import parse, eval_lisp, standard_env
-from lispfun.run import load_eval, run_file
+from run_hosted import load_eval, run_file
 
 SUPPORT = os.path.join(os.path.dirname(__file__), "lisp", "import_support.lisp")
 MAIN = os.path.join(os.path.dirname(__file__), "lisp", "import_main.lisp")
