+++ lispfun/hosted/README.md
@@ -1,5 +1,5 @@
 # Self-hosted Evaluator
 
 This directory contains the Lisp source files for the `eval2` evaluator.  The
-Python runner `lispfun/run.py` loads `evaluator.lisp` from here and executes
-expressions via `(eval2 ...)`.
+Python runner ``run_hosted.py`` loads ``evaluator.lisp`` from here and executes
+expressions via ``(eval2 ...)``.
